Fix COPYDATASTRUCT for x64 target (fix exception in SingleInstanceApp::NotifyPreviousInstance)

+++ PluginCore/PluginCore/Win32.cs
@@ -63,7 +63,7 @@ namespace PluginCore
 
         public struct COPYDATASTRUCT
         {
-            public Int32 dwData;
+            public IntPtr dwData;
             public Int32 cbData;
             public IntPtr lpData;
         }
